requirements.txt
pandas
numpy
streamlit
sentence-transformers
faiss-cpu
transformers
torch
